move direct named tests to their own file
